requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
gunicorn==23.0.0
numpy==2.1.3
pillow==11.1.0
python-multipart==0.0.20
